requirements.txt
django
djangorestframework
django-cors-headers
python-decouple
psycopg-binary
psycopg
nodeenv



# npm install react-router-dom
# npm install axios
# pip install django psycopg2